[previews] Rename RawVideoPlayer to MultiVideoViewer
The component is a multi-source viewer (videos + pictures stitched with
framesPerImage), not a "raw" anything, and it is consumed by more than
the playlists page. Rename it and move it to src/components/previews
alongside the other viewers (PictureViewer, ObjectViewer, SoundViewer,
PdfViewer). Internal name attribute, imports and template tags follow.

The new name also makes the upcoming PlaylistPlayer composition-API
refactor easier to read: the multi-entity playback responsibility is
now visible in the component name itself.
