+++ src/components/pages/playlists/PlaylistPlayer.vue
@@ -1057,7 +1057,7 @@ import ObjectViewer from '@/components/previews/ObjectViewer.vue'
 import PencilPicker from '@/components/widgets/PencilPicker.vue'
 import PlaylistedEntity from '@/components/pages/playlists/PlaylistedEntity.vue'
 import PictureViewer from '@/components/previews/PictureViewer.vue'
-import MultiVideoViewer from '@/components/pages/playlists/MultiVideoViewer.vue'
+import MultiVideoViewer from '@/components/previews/MultiVideoViewer.vue'
 import PdfViewer from '@/components/previews/PdfViewer.vue'
 import PreviewRoom from '@/components/widgets/PreviewRoom.vue'
 import SelectTaskTypeModal from '@/components/modals/SelectTaskTypeModal.vue'
